Struct chipo::Chip8

pub struct Chip8 { /* fields omitted */ }

A chip-8 emulator.

Methods

impl Chip8

pub fn new() -> Self

Constructs a new Chip8.

pub fn load_rom(&mut self, rom: &[u8], should_wrap: bool)

Loads a rom and sets the drawing mode of the emulator. If should_wrap is true, then all pixels drawn outside of the drawable area will wrap to the other side, else they will be ignored.

pub fn execute_cycle(&mut self)

Runs one fetch-decode-execute cycle.

pub fn decrement_timers(&mut self)

Decrement the delay and sound timer by one tick.

pub fn screen(&self) -> *const u8

Returns a pointer to a byte array that represents the screen. The screen will have screen_width * screen_height / 8 bytes in row-major order. Each byte represents 8 bits in little-endian. 1 represents that the pixel is black, while 0 represents that the pixel is white.

pub fn screen_width(&self) -> usize

Returns the width of the screen in pixels.

pub fn screen_height(&self) -> usize

Returns the height of the screen in pixels.

pub fn press_key(&mut self, index: usize)

Sets the state of a key to be pressed. index is the index of the key in row-major order.

pub fn release_key(&mut self, index: usize)

Sets the state of a key to be released. index is the index of the key in row-major order.

pub fn should_draw(&self) -> bool

Returns true if the screen has been updated and should be redrawn.

pub fn should_beep(&self) -> bool

Returns true if the a beep should be made.

pub fn program_counter(&self) -> u16

Returns the value of the program counter register.

pub fn index(&self) -> u16

Returns the value of the index register.

pub fn registers(&self) -> *const u8

Returns a pointer to a byte array that represents the 16 data registers.